Refactor out seperate help utils for InlineHelp and generic Support Search
